+++ graphene_gae/webapp2/__init__.py
@@ -1,7 +1,9 @@
 import logging
 import json
 import webapp2
+from webapp2_extras import jinja2
 import six
+
 
 from graphql import GraphQLError, format_error as format_graphql_error
 
@@ -128,6 +130,151 @@ class GraphQLHandler(webapp2.RequestHandler):
         self.response.out.write(serialized_data)
 
 
+class GraphiqlHandler(webapp2.RequestHandler):
+    GRAPHIQL_VERSION = '0.7.1'
+
+    GRAPHIQL_TEMPLATE = '''
+    The request to this GraphQL server provided the header "Accept: text/html"
+    and as a result has been presented GraphiQL - an in-browser IDE for
+    exploring GraphQL.
+    If you wish to receive JSON, provide the header "Accept: application/json" or
+    add "&raw" to the end of the URL within a browser.
+    -->
+    <!DOCTYPE html>
+    <html>
+    <head>
+      <style>
+        html, body {
+          height: 100%;
+          margin: 0;
+          overflow: hidden;
+          width: 100%;
+        }
+      </style>
+      <meta name="referrer" content="no-referrer">
+      <link href="//cdn.jsdelivr.net/graphiql/{{graphiql_version}}/graphiql.css" rel="stylesheet" />
+      <script src="//cdn.jsdelivr.net/fetch/0.9.0/fetch.min.js"></script>
+      <script src="//cdn.jsdelivr.net/react/15.0.0/react.min.js"></script>
+      <script src="//cdn.jsdelivr.net/react/15.0.0/react-dom.min.js"></script>
+      <script src="//cdn.jsdelivr.net/graphiql/{{graphiql_version}}/graphiql.min.js"></script>
+    </head>
+    <body>
+      <script>
+        // Collect the URL parameters
+        var parameters = {};
+        window.location.search.substr(1).split('&').forEach(function (entry) {
+          var eq = entry.indexOf('=');
+          if (eq >= 0) {
+            parameters[decodeURIComponent(entry.slice(0, eq))] =
+              decodeURIComponent(entry.slice(eq + 1));
+          }
+        });
+        // Produce a Location query string from a parameter object.
+        function locationQuery(params) {
+          return '?' + Object.keys(params).map(function (key) {
+            return encodeURIComponent(key) + '=' +
+              encodeURIComponent(params[key]);
+          }).join('&');
+        }
+        // Derive a fetch URL from the current URL, sans the GraphQL parameters.
+        var graphqlParamNames = {
+          query: true,
+          variables: true,
+          operationName: true
+        };
+        var otherParams = {};
+        for (var k in parameters) {
+          if (parameters.hasOwnProperty(k) && graphqlParamNames[k] !== true) {
+            otherParams[k] = parameters[k];
+          }
+        }
+        var fetchURL = locationQuery(otherParams);
+        // Defines a GraphQL fetcher using the fetch API.
+        function graphQLFetcher(graphQLParams) {
+          return fetch(fetchURL, {
+            method: 'post',
+            headers: {
+              'Accept': 'application/json',
+              'Content-Type': 'application/json'
+            },
+            body: JSON.stringify(graphQLParams),
+            credentials: 'include',
+          }).then(function (response) {
+            return response.text();
+          }).then(function (responseBody) {
+            try {
+              return JSON.parse(responseBody);
+            } catch (error) {
+              return responseBody;
+            }
+          });
+        }
+        // When the query and variables string is edited, update the URL bar so
+        // that it can be easily shared.
+        function onEditQuery(newQuery) {
+          parameters.query = newQuery;
+          updateURL();
+        }
+        function onEditVariables(newVariables) {
+          parameters.variables = newVariables;
+          updateURL();
+        }
+        function onEditOperationName(newOperationName) {
+          parameters.operationName = newOperationName;
+          updateURL();
+        }
+        function updateURL() {
+          history.replaceState(null, null, locationQuery(parameters));
+        }
+        // Render <GraphiQL /> into the body.
+        ReactDOM.render(
+          React.createElement(GraphiQL, {
+            fetcher: graphQLFetcher,
+            onEditQuery: onEditQuery,
+            onEditVariables: onEditVariables,
+            onEditOperationName: onEditOperationName,
+            query: {{ query|tojson }},
+            response: {{ result|tojson }},
+            variables: {{ variables|tojson }},
+            operationName: {{ operation_name|tojson }},
+          }),
+          document.body
+        );
+      </script>
+    </body>
+    </html>
+    '''
+
+    @webapp2.cached_property
+    def jinja2(self):
+        # Returns a Jinja2 renderer cached in the app registry.
+        return jinja2.get_jinja2()
+
+    def __render_graphiql(self, graphiql_version=None, graphiql_template=None, **kwargs):
+        graphiql_version = graphiql_version or self.GRAPHIQL_VERSION
+        graphiql_template = graphiql_template or self.GRAPHIQL_TEMPLATE
+
+        template = self.jinja2.environment.from_string(graphiql_template)
+        return template.render(graphiql_version=graphiql_version, **kwargs)
+
+    def get(self):
+        query, operation_name, variables, result = self._get_grapl_params()
+        vars = dict(query=query, operation_name=operation_name, variables=variables, result=result)
+
+        self.response.write(self.__render_graphiql(**vars))
+
+
+    def _get_grapl_params(self):
+        request_data = self.request.GET
+
+        query = request_data.get('query', '')
+        operation_name = request_data.get('operation_name', '')
+        variables = request_data.get('variables', '')
+        result = request_data.get('result', '')
+        return query, operation_name, variables, result
+
+
 graphql_application = webapp2.WSGIApplication([
-    ('/graphql', GraphQLHandler)
+    ('/graphql', GraphQLHandler),
+    ('/graphiql', GraphiqlHandler),
 ])